pyproject.toml
[project]
name = "alpaca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.39.4",
    "keyring>=25.6.0",
]
